Add files via upload
Here the slides of the presentation I gave for the social axe meeting. 
The story I tried to tell it's simple: Github is an easy-to-learn tool that can make your work more productive, agile and that can avoid big wasting of time.
I provided some basic function of GitHub (mostly for working with a script) and I provided a concrete example of how GitHub can make the realization of a project much faster and more efficient.